requirements.txt
# main
flask==2.2.2
Flask-WTF==1.1.1
email_validator==1.3.1
requests==2.28.1
psycopg2-binary==2.9.5
Flask-SQLAlchemy==3.0.3

# linters
black==22.6.0

# testing
blinker==1.5

# google drive
google-api-python-client==2.80.0
google-auth-httplib2
google-auth-oauthlib==1.0.0

# development
python-dotenv==1.0.0